+++ engine/src/core_data/src/numerics.rs
@@ -28,7 +28,7 @@ use serde::{Deserialize, Serialize};
     Serialize,
     Deserialize,
 )]
-pub struct Energy(pub u64);
+pub struct Energy(pub u32);
 
 /// Represents the 'power' of characters; the primary way in which players earn
 /// victory points.
@@ -56,7 +56,7 @@ pub struct Energy(pub u64);
     Serialize,
     Deserialize,
 )]
-pub struct Spark(pub u64);
+pub struct Spark(pub u32);
 
 /// Victory points. Enable the player to win the game.
 #[derive(
@@ -83,4 +83,4 @@ pub struct Spark(pub u64);
     Serialize,
     Deserialize,
 )]
-pub struct Points(pub u64);
+pub struct Points(pub u32);
